backend/requirements.txt
annotated-types
anyio
black
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
cryptography
dnspython
ecdsa
email-validator
fastapi
feedparser
flake8
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
iniconfig
isort
jmespath
jq
markdown-it-py
mccabe
mdurl
motor
mypy
mypy_extensions
numpy
oauthlib
packaging
pandas
passlib
pathspec
platformdirs
pluggy
proto-plus
protobuf
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
Pygments
PyJWT
pymongo
pyparsing
pytest
python-dateutil
python-dotenv
python-jose
python-multipart
pytokens
pytz
requests
requests-oauthlib
rich
rsa
s3transfer
s5cmd
sgmllib3k
shellingham
six
sniffio
starlette
tqdm
typer
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchfiles
